requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.2.0
awscli==1.33.21
awsebcli==3.20.10
blinker==1.8.2
boto3==1.34.139
botocore==1.34.139
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cement==2.8.2
certifi @ file:///C:/b/abs_55jxithrm3/croot/certifi_1717618077850/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chex==0.1.8
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.3
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
docker==7.1.0
docutils==0.16
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
gast==0.6.0
gevent==24.2.1
geventhttpclient==2.0.2
gmpy2 @ file:///C:/ci_311/gmpy2_1677743390134/work
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h5py==3.11.0
humanfriendly==10.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.9.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_311/mkl_random_1676481991689/work
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy @ file:///C:/b/abs_54abayvc9j/croot/numpy_and_numpy_base_1682520598361/work
onnx==1.16.1
onnxruntime==1.18.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging @ file:///C:/b/abs_cc1h2xfosn/croot/packaging_1710807447479/work
pandas==2.2.2
pathos==0.3.2
pathspec==0.10.1
pillow @ file:///C:/b/abs_95ls_qh4c2/croot/pillow_1714398870615/work
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
pooch @ file:///C:/b/abs_a8nmng7d_x/croot/pooch_1695850149827/work
pox==0.3.4
ppft==1.7.6.8
protobuf==3.20.3
psutil==6.0.0
pyasn1==0.6.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
pyreadline3==3.4.1
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-rapidjson==1.17
pytz==2024.1
pywin32==306
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
referencing==0.35.1
requests @ file:///C:/b/abs_9c5n24p1y7/croot/requests_1716902867639/work
rich==13.7.1
rpds-py==0.18.1
rsa==4.7.2
s3transfer==0.10.2
sagemaker==2.224.4
schema==0.7.7
scipy==1.10.1
semantic-version==2.8.5
six @ file:///tmp/build/80754af9/six_1644875935023/work
smdebug-rulesconfig==1.0.1
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tblib==3.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
tf2onnx==1.16.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
tritonclient==2.46.0
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.1.9
Werkzeug==3.0.3
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
